Removed un-needed parameters `self._t0` and `self._tf`.
`self._t0` and `self._tf` refereed to the start and end times
of the signal under analysis. These parameters were previously
used to set the x-extent of the spectrogram plot. However,
the spectral slices returned by `mlab.specgram(...)` actually
belong to the times at the *midpoint* of the window.
Thus, the minimum time in the spectrogram plot should be

            tmin = t0 + ((NFFT / 2) / Fs)

where `t0` is the start time of the signal, `NFFT` is the
number of data points per window, and `Fs` is the sampling
frequency. Now, `mlab.specgram(...)` already returns times

            t_j = ((NFFT / 2) / Fs) * j

for the jth window. We simply add `t0` to the time array
to obtain the correct value for `tmin`. Similarly,
this provides the correct value for `tmax`.

Thus, we no longer need to carry around the unused
(and potentially) confusing `self._t0` and `self._tf`.