Revamping spectral density estimation class.
The new class will allow for computation of autospectral densities,
cross-spectral densities, phase angles, and coherences, as well
as easy tools for visualizing the data.

The current commit implements identification of the type of
spectral density (either "autospectral" or "cross-spectral")
and tests to ensure the routines are working as intended.