Naively detrending can introduce artifacts into computed spectrum.
Even something as simple as `mean` or `linear` detrending can introduce
artifacts into the computed spectrum; for example, computing the
autospectral density of signal `x` as

    asd1, f1 = mlab.psd(x, detrend=None)
    asd2, f2 = mlab.csd(x, x.copy(), detrend=None)

gives the expected result that

    np.alltrue(asd1 == asd2)

is `True`. However, if linear detrending is instead performed

    asd1_lin, f1 = mlab.psd(x, detrend='linear')
    asd2_lin, f2 = mlab.csd(x, x.copy(), detrend='linear')

then the expected equality no longer holds, and there may be
relative differences in the computed spectra ~10-20% - bad!!!
Note that this discrepancy holds for the `mlab` spectral routines
and is *not* a result of my wrapper around them.

As a result, I've modified the default settings of the spectral objects
to have `None` detrending to avoid this unexpected behavior.