Corrected phase angle plotting in CrossSpectralDensity class.
The previous implementation suffered from an off-by-one error
(one too many bins) that caused non-trivial phase offsets for
large phase angle. The previous off-by-one error resulted from
the failure to realize that the (default) domain should be
[-pi, pi) rather than [-pi, pi]; this follows from the fact
that -pi and pi are in fact the same point on the unit circle,
so they should *not* have distinct bins.

In addition to correcting the mapping of phase angle to color bin,
the phase angle plotting routine allows arbitrary "wrapping"
of phase angles. That is, if we do not know a mode's
rotation direction, we would have to use a range [-pi, pi);
however, if we know the mode is rotating in the e.g. *positive*
direction, we can specify the domain as [0, 2 * pi), thereby
doubling the resolution of our computation.